apps/backend/requirements.txt
asgiref==3.7.2
Django==4.2.8
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
